+++ R/aaa.R
@@ -1,5 +1,6 @@
 
 #' @importFrom tools R_user_dir
+#' @importFrom utils str
 .onAttach <- function(libname, pkgname) {
   plantumlOptions(
     jar_name = file.path( tools::R_user_dir(package = "plantuml", "cache"), "jar", "plantuml.jar"),
@@ -10,4 +11,10 @@
     server_port = 8765
   )
   plantuml_knit_engine_register()
+  server_start()
+  si <- server_info()
+  msg <- paste(names(si), ": ", unlist(si), collapse = "\n")
+  packageStartupMessage(utils::str(si))
+  packageStartupMessage(" You have to stop the server manually by calling `server_stop()` before quitting R!\n")
 }
+