+++ doc/use_with_knitr.R
@@ -1,9 +1,0 @@
-## ---- include = FALSE---------------------------------------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-
-## ----setup, include=TRUE------------------------------------------------------
-plantuml::plantuml_knit_engine_register()
-